Created separate folder for Overview component.
