+++ src/components/ProductDetails.js
@@ -3,7 +3,7 @@ import { useParams } from '@reach/router';
 import { Layout, Space, Divider } from 'antd';
 import { UserOutlined, LaptopOutlined, NotificationOutlined } from '@ant-design/icons';
 import { Input } from 'antd';
-import Overview from './Overview';
+import Overview from './OverviewComponent/Overview';
 import RelatedItems from './RelatedItems';
 import QA from './QA';
 import Ratings from './Ratings';
