+++ src/index.css
@@ -39,7 +39,7 @@ code {
 }
 
 .carouselPlaceHolder {
-  width: 800px;
+  width: 100%;
   height: 600px;
 }
 
@@ -54,7 +54,7 @@ code {
 
 .outer-carousel {
   position: absolute;
-  width: 800px;
+  width: 100%;
   height: 600px;
   z-index: 9999;
   transition: width 2s;
@@ -75,7 +75,7 @@ code {
 .expand-button {
   position: relative;
   top: 20px;
-  right: -780px;
+  right: -675px;
   z-index: 99999;
 }
 
@@ -91,12 +91,12 @@ code {
 
 .arrows.right {
   top: -330px;
-  left: 390px;
+  left: 340px;
 }
 
 .arrows.left {
   top: -300px;
-  left: -390px
+  left: -340px
 }
 
 .small-image {
